refactor: Moving call logic to ProcessBase using back calls to ProcessNode
Enables a more consistent interception of the call logic, e.g. for caching operations and parameter access at every step of the process chain.
